interp: snap the Python-visible frame offsets off a CALL's cache words
`last_instr` is a dispatch coordinate: the interpreter stores the unit it is
about to run, pyre materializes each cache word as an `Instruction::Cache` it
dispatches as a no-op, and every resume writer stores `pc - 1` against
`next_instr() == last_instr + 1`.  A caller suspended at a `CALL` therefore
carries that call's LAST CACHE WORD, and the three readers that hand the field
to Python handed that word out: `f_lasti` reported 58 where `dis` places the
`CALL` at 52 and the next instruction at 60, `tb_lasti` minted the same word,
and `frame.f_lineno = N` indexed `mark_stacks`, whose cache entries keep
`MARK_UNINITIALIZED`, so it answered "can't jump from unreachable code".

`pyopcode::owning_instruction` walks a cache word back to the opcode that owns
it -- the backward mirror of `skip_caches` -- and is the identity on any value
that already names an instruction.  `PyFrame::executing_instr` applies it
behind `fget_f_lasti`, `record_application_traceback` applies it to the
`tb_lasti` mint, and `fset_f_lineno` indexes with it.  `next_instr()` and the
resume consumers keep reading the field raw, so the coordinate they round-trip
is untouched and the compiled channel
(`try_walker_specialize_frame_lasti` -> `walker_prove_owned_frame_pc`) already
emitted an instruction.

`caller_f_lasti_across_residual_call` moves out of `_pending`: it read
`[(52, 11074), (58, 8926)]` and now reads `[(52, 20000)]` on dynasm and
cranelift alike, matching cpython.
